requirements.txt
flask
flask-cors
python-dotenv
requests
firebase-admin



# venv\Scripts\activate